requirements.txt
streamlit==1.9
snowflake-connector-python==2.7.7
streamlit-folium
folium